move tests to test/it/serde.rs
